[FIX / RENAME] (User) 유저 이모지 조회 수정, 파일 패키지 이동
