+++ src/main/java/com/example/lastproject/domain/user/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.example.lastproject.domain.user.controller;
 
-import com.example.lastproject.domain.auth.entity.AuthUser;
+import com.example.lastproject.common.dto.AuthUser;
 import com.example.lastproject.domain.user.dto.request.UserChangePasswordRequest;
 import com.example.lastproject.domain.user.dto.request.UserUpdateRequest;
 import com.example.lastproject.domain.user.dto.response.UserResponse;
